requirements.txt
aniso8601==7.0.0
bson==0.5.10
git+https://github.com/leoalvs/buslane.git
click==7.1.2
fastapi==0.63.0
h11==0.12.0
injector==0.17.0
promise==2.3
pydantic==1.7.3
mongoengine==0.22.1
pymongo==3.11.2
python-dateutil==2.8.1
Rx==1.6.1
six==1.15.0
starlette==0.13.6
stopwatch.py==1.0.1
typing-extensions==3.7.4.3
uvicorn==0.13.3
python-stdnum==1.15
factory-boy==2.12.0
freezegun==1.1.0
object-mapper==1.0.7
pytest==6.2.2
mongomock==3.22.1
coverage==5.4